Oracle: Fix detection of :decimal vs :integer vs :boolean columns.
RubyJdbcConnection had some code in it that would cause typeFromResultSet to
return "NUMBER" for both types of columns, instead of something like
"NUMBER(38,2)".

Oracle's default NUMBER has a NULL DECIMAL_DIGITS, and this must be expressed
as "NUMBER", not eg. "NUMBER(22)" (the latter is an integer type), so we
special-case it.

For reference: this is for "Oracle Database 11g Release 11.2.0.1.0 - 64bit Production"